Funcionalidad al 95%, estilos al 10%
